requirements.txt
--index-url https://pypi.org/simple/
--extra-index-url https://test.pypi.org/simple/
--prefer-binary

# Core libraries
PyYAML # <--Make sure at least this is installed in local dev
alembic
annotated-types
anyio
asyncio
certifi
cffi
click
cryptography
exceptiongroup
fastapi~=0.110.0
greenlet
h11
httpcore
httpx
idna
iniconfig
jinja2
mypy
mypy-extensions
numpy
openai
packaging
pdfplumber~=0.11.6
pip
pluggy
pycparser
pydantic~=2.10.3
pydantic_core
PyMySQL
pytest>=7.0.0
python-dotenv
python-multipart
qdrant-client~=1.13.3
requests
sentence-transformers~=4.0.1
setuptools~=75.8.0
sniffio
SQLAlchemy
sse_starlette
starlette
sympy
tenacity
together
torch
typing
typing_extensions
uvicorn[standard]
validators~=0.34.0
websockets

# Azure-specific inference
azure-core
azure-ai-inference

# google

google-generativeai

# Ollama integration
ollama
#
groq

# PySMB (Samba file access)
pysmb

# api-keys
passlib[bcrypt]

# Internal packages
projectdavid==1.17.0
projectdavid-common==0.12.0

# From Test PyPI (only these should resolve there)


